requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.28.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
transformers==4.40.0
numpy==1.26.4
einops==0.7.0
pillow==10.3.0
timm==1.0.7
tokenizers==0.19.1
triton==2.0.0
typing_extensions==4.11.0
torchsde==0.2.6
tqdm==4.66.2
pytorch-lightning==2.2.2
jsonargparse==4.28.0
matplotlib==3.8.4
omegaconf==2.3.0
typeshed_client==2.5.1
docstring_parser==0.16
kornia==0.7.2
open-clip-torch==2.24.0
xformers==0.0.20
opencv-python==4.10.0.84
accelerate==0.29.3
gdown==5.2.0